+++ src/hooks/ContextProvider.js
@@ -1,4 +1,4 @@
-import React, { createContext, useCallback, useContext, useEffect, useState } from 'react';
+import React, { createContext, useCallback, useContext, useState } from 'react';
 import contentfulService from 'services/ContentfulService';
 
 export const AppContext = createContext();
@@ -6,31 +6,10 @@ export const AppContext = createContext();
 export const AppContextProvider = (props) => {
     const [audios, setAudios] = useState([]);
     const [imageToOpen, setImageToOpen] = useState(null);
-    const [isAppNameVisible, setIsAppNameVisible] = useState(false);
-    const [isBurguerMenuInverted, setIsBurguerMenuInverted] = useState(false);
-    const [isDesktopMenuOpen, setIsDesktopMenuOpen] = useState(false);
-    const [isHeaderInverted, setIsHeaderInverted] = useState(false);
-    const [isTitleInverted, setIsTitleInverted] = useState(false);
     const [isLoadingAudios, setIsLoadingAudios] = useState(false);
     const [isLoadingPosts, setIsLoadingPosts] = useState(false);
     const [isMenuOpen, setIsMenuOpen] = useState(false);
-    const [isTabOpen, setIsTabOpen] = useState(false);
     const [posts, setPosts] = useState([]);
-
-    useEffect(() => {
-        if (isHeaderInverted) {
-            setIsBurguerMenuInverted(true);
-            setIsTitleInverted(true);
-        } else {
-            setIsBurguerMenuInverted(false);
-            setIsTitleInverted(false);
-        }
-    }, [isHeaderInverted]);
-
-    useEffect(() => {
-        if (isBurguerMenuInverted && isTitleInverted) setIsHeaderInverted(true);
-        if (!isBurguerMenuInverted && !isTitleInverted) setIsHeaderInverted(false);
-    }, [isBurguerMenuInverted, isTitleInverted]);
 
     const getVideos = useCallback(async () => {
         setIsLoadingPosts(true);
@@ -61,24 +40,12 @@ export const AppContextProvider = (props) => {
         getVideos,
         getVideosById,
         imageToOpen,
-        isAppNameVisible,
-        isBurguerMenuInverted,
-        isDesktopMenuOpen,
-        isHeaderInverted,
         isLoadingAudios,
         isLoadingPosts,
         isMenuOpen,
-        isTabOpen,
-        isTitleInverted,
         posts,
         setImageToOpen,
-        setIsAppNameVisible,
-        setIsBurguerMenuInverted,
-        setIsDesktopMenuOpen,
-        setIsHeaderInverted,
-        setIsMenuOpen,
-        setIsTabOpen,
-        setIsTitleInverted
+        setIsMenuOpen
     };
 
     return <AppContext.Provider {...props} value={value} />;