+++ src/pages/home/music/MusicSection.js
@@ -57,11 +57,13 @@ const MusicSection = ({ gsap, ...props }) => {
         gsap.fromTo(
             '#musica',
             {
-                opacity: 1
+                opacity: 1,
+                yPercent: 0
             },
             {
                 opacity: 0,
-                duration: 0.2,
+                yPercent: -100,
+                duration: 0.01,
                 scrollTrigger: {
                     scroller: '.main-page',
                     trigger: '#musica',